+++ packages/app/components/stack-modal/StackModal.tsx
@@ -23,6 +23,7 @@ import {
 import {
   formatFrequencyHours,
   formatTimestampToDateWithTime,
+} from "@/utils/datetime";
 import { FromToStackTokenPair } from "@/components/FromToStackTokenPair";
 import { calculateAveragePrice } from "@/components/StacksTable";
